Rename useSearchSpotifyMusic to useSearchSpotifyTrack
